[SR-1807] Diagnose primitive reference cycles
Code is pretty much cribbed directly from Clang.  Here we try to use
the same walker that is already traversing Expr's looking for invalid
uses of 'self' to catch those same uses being captured in primitive
reference cycles.  Our analysis looks for more than just 'self', though.

Analysis is limited to assignment expressions that use the variable
inside the closure expression that forms their initializer and closures
in argument position.  By design, this will not catch everything.

Note that this currently does not handle the case where we are inside
a pattern binding for a lazy var with a block in its initializer.  Those
can still capture self.

I am worried about the performance impact of having to compute capture
lists for every closure we come across that might happen to be
a referrant of an expression that has reference semantics.  That is
quickly going to get expensive.